Add ability to authenticate with header
Example:
```sh
curl -H "Authorization: 00000000-0000-0000-0000-000000000000" http://0.0.0.0:5000/admin/token?room=test_room
```
