+++ app/models/user.py
@@ -1,15 +1,17 @@
 from flask import request
 from flask_login import current_user
 
-from .. import db, socketio
+from .. import db, socketio, login_manager
 
 from . import Base, user_room
+
+from .token import Token
 
 
 class User(Base):
     __tablename__ = 'User'
 
-    name = db.Column(db.String, nullable=False)
+    name = db.Column(db.String)
     token = db.relationship("Token", backref="user", uselist=False)
     rooms = db.relationship("Room", secondary=user_room, back_populates="users", lazy='dynamic')
     session_id = db.Column(db.String, unique=True)
@@ -36,6 +38,86 @@ class User(Base):
 
     def get_id(self):
         return self.id
+
+
+@login_manager.user_loader
+def load_user(id):
+    return User.query.get(int(id))
+
+
+@login_manager.request_loader
+def load_user_from_request(request):
+    token = None
+    token_id = request.headers.get('Authorization')
+    if token_id:
+        token = Token.query.get(token_id)
+    if not token:
+        token_id = request.args.get('token')
+        if token_id:
+            token = Token.query.get(token_id)
+
+    if token:
+        if not token.user:
+            token.user = User(name=request.args.get('name'))
+            db.session.commit()
+        return token.user
+    return None
+
+
+@socketio.on('get_rooms_by_user')
+def _get_rooms_by_user(id):
+    if not current_user.get_id():
+        return False, "invalid session id"
+    if id and not (current_user.token.permissions.query_room and current_user.token.permissions.query_user):
+        return False, "insufficient rights"
+
+    if id:
+        user = User.query.get(id)
+    else:
+        user = current_user
+
+    if user:
+        return True, [room.as_dict() for room in user.rooms]
+    else:
+        return False, "user does not exist"
+
+
+@socketio.on('get_permissions_by_user')
+def _get_permissions_by_user(id):
+    if not current_user.get_id():
+        return False, "invalid session id"
+    if id and not (current_user.token.permissions.query_permissions and current_user.token.permissions.query_user):
+        return False, "insufficient rights"
+
+    if id:
+        user = User.query.get(id)
+    else:
+        user = current_user
+
+    if user:
+        return True, user.token.permissions.as_dict()
+    else:
+        return False, "user does not exist"
+
+
+@socketio.on('get_layouts_by_user')
+def _get_layouts_by_user(id):
+    if not current_user.get_id():
+        return False, "invalid session id"
+
+    if id:
+        if not (current_user.token.permissions.query_layout and
+                current_user.token.permissions.query_room and
+                current_user.token.permissions.query_user):
+            return False, "insufficient rights"
+        user = User.query.get(id)
+    else:
+        user = current_user
+
+    if user:
+        return True, {room.name: room.layout.as_dict() for room in user.rooms}
+    else:
+        return False, "user does not exist"
 
 
 @socketio.on('get_user')
